ml/requirements.txt
# Database access
psycopg2-binary==2.9.9

# Data processing
pandas==2.2.0
numpy==1.26.3

# Optional: for Parquet export
pyarrow==15.0.0

# Optional: for Jupyter notebooks
jupyter==1.0.0
ipykernel==6.29.0

# ML libraries (add as needed)
scikit-learn==1.4.0

# Airflow for ETL pipelines
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.10.0

# For AWS Lambda (optional)
requests==2.31.0
